move MPL_COMPAT to util re #291
