+++ mslice/models/cut/matplotlib_cut_plotter.py
@@ -1,6 +1,6 @@
 from __future__ import (absolute_import, division, print_function)
 import mslice.plotting.pyplot as plt
-from mslice.app import MPL_COMPAT
+from mslice.util import MPL_COMPAT
 from .cut_plotter import CutPlotter
 from .mantid_cut_algorithm import output_workspace_name
